requirements.txt
crewai-tools[mcp]
python-dotenv
requests
mcpadapt
google-genai
streamlit
pathlib